refactor: create and use `useGoogleDomain`
